(lpdoc) Relax no_propuses option (at least point to def)
Previously this option excluded property uses from the indices, which
was needed for very large documents and texinfo.

In principle this is not needed for HTML but we still use it on the
premise that indices could be huge.

This commit introduces a 'use_noidx' mode to the idx_env internal
formatting command, which is exactly like the 'use' mode but is not
included in the 'use' index. The behaviour for texinfo backend is the
same as before.

Note that better text index representation (e.g. trigram based) could
make this irrelevant.

Src-commit: 731154f829f1b57981f6cf7e2fba77fd10d17c34