(lpdoc) Added "is_operational" checks for lpdoc backends
This check is used to disable non-operational backend formats due to
missing dependencies (e.g., tex, makeinfo, etc.). For example, if 'tex'
is missing but 'makeinfo' is available, only 'info' documentation will
be generated.

Error messages about non-operational backends formats are shown only
when the format is specified explicitly. When using the 'all' formats,
only the operational formats are selected.

Src-commit: 12d52471e50edc0f39cccabe06f741e8c95018ed